comp driving

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutonMethods.java
@@ -117,7 +117,7 @@ public class AutonMethods {
 
     }
 
-    public long maps(long x, long in_min, long in_max, long out_min, long out_max) {
+    public double maps(double x, double in_min, double in_max, double out_min, double out_max) {
         return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
     }
